third_party/blink/web_tests/external/wpt/svg/styling/presentation-attributes-special-cases-expected.txt

This is a testharness.js-based test.
[FAIL] x, y, width, and height presentation attributes supported on svg element
  assert_true: Presentation attribute width="1" should be supported on svg element expected true got false
[FAIL] x, y, width, and height presentation attributes supported on symbol element
  assert_true: Presentation attribute x="1" should be supported on symbol element expected true got false
[FAIL] x, y, width, and height presentation attributes supported on use element
  assert_true: Presentation attribute width="1" should be supported on use element expected true got false
[FAIL] fill presentation attribute not supported on discard
  assert_false: Presentation attribute fill="blue" should not be supported on discard element expected false got true
Harness: the test ran to completion.
